requirements.txt
pandas
PySide6
matplotlib
Pillow
reportlab
python-dotenv
requests